Add `MoxieAssetPlugin`
